requirements.txt
fastapi==0.112.2
sqlalchemy==2.0.32
starlette==0.38.2
Jinja2==3.1.4
pydantic==2.8.2
uvicorn==0.30.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
databases==0.9.0
pydantic==2.8.2
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic[email]==2.8.2